TextClassification

TextClassification.MicroTC

StatsBase.predict

TextClassification.create

TextSearch.vectorize

TextClassification.MicroTC — Method

MicroTC(config::MicroTC_Config, train_corpus::AbstractVector, train_y::CategoricalArray; verbose=true)
MicroTC(config::MicroTC_Config, textmodel::TextModel, train_X::AbstractVector{S}, train_y::CategoricalArray; verbose=true) where {S<:SVEC}

Creates a MicroTC model on the given dataset and configuration

StatsBase.predict — Method

predict(tc::MicroTC, text)
predict(tc::MicroTC, vec::SVEC)

Predicts the label of the given input

TextClassification.create — Method

create(config, train_X, train_y) # config describes a text model
create(config, train_X, train_y, dim) # config describes a classifier

Creates a new object from a configuration and a train / test datasets.

train_X is an array of BOW objects for text models

train_X is an array of SVEC objects for classifiers

TextSearch.vectorize — Method

vectorize(tc::MicroTC, text)
vectorize(tc::MicroTC, bow::BOW)

Creates a weighted vector using the model. The input text can be a string or an array of strings; it also can be an already computed bag of words.